+++ apps/docs/content/components/date-range-picker/index.ts
@@ -1,0 +1,47 @@
+import usage from "./usage";
+import disabled from "./disabled";
+import readOnly from "./readonly";
+import required from "./required";
+import variants from "./variants";
+import labelPlacements from "./label-placements";
+import description from "./description";
+import errorMessage from "./error-message";
+import errorMessageFunction from "./error-message-function";
+import withTimeField from "./with-time-field";
+import selectorIcon from "./selector-icon";
+import controlled from "./controlled";
+import timeZones from "./time-zones";
+import granularity from "./granularity";
+import minAndMaxDate from "./min-and-max-date";
+import internationalCalendar from "./international-calendar";
+import unavailableDates from "./unavailable-dates";
+import visibleMonth from "./visible-month";
+import pageBehavior from "./page-behavior";
+import nonContigous from "./non-contiguous";
+import presets from "./presets";
+import customStyles from "./custom-styles";
+
+export const dateRangePickerContent = {
+  usage,
+  disabled,
+  readOnly,
+  required,
+  variants,
+  labelPlacements,
+  description,
+  errorMessage,
+  errorMessageFunction,
+  withTimeField,
+  selectorIcon,
+  controlled,
+  timeZones,
+  granularity,
+  minAndMaxDate,
+  internationalCalendar,
+  unavailableDates,
+  visibleMonth,
+  pageBehavior,
+  nonContigous,
+  presets,
+  customStyles,
+};