+++ packages/components/autocomplete/stories/autocomplete.stories.tsx
@@ -1,5 +1,9 @@
+import type {ValidationResult} from "@react-types/shared";
+
 import React, {Key} from "react";
 import {Meta} from "@storybook/react";
+import {useForm} from "react-hook-form";
+import {useFilter} from "@react-aria/i18n";
 import {autocomplete, input, button} from "@nextui-org/theme";
 import {
   Pokemon,
@@ -9,6 +13,7 @@ import {
   Animal,
   User,
 } from "@nextui-org/stories-utils";
+import {useAsyncList} from "@react-stately/data";
 import {useInfiniteScroll} from "@nextui-org/use-infinite-scroll";
 import {PetBoldIcon, SearchLinearIcon, SelectorIcon} from "@nextui-org/shared-icons";
 import {Avatar} from "@nextui-org/avatar";
@@ -60,6 +65,12 @@ export default {
         type: "boolean",
       },
     },
+    validationBehavior: {
+      control: {
+        type: "select",
+      },
+      options: ["aria", "native"],
+    },
   },
   decorators: [
     (Story) => (
@@ -69,6 +80,13 @@ export default {
     ),
   ],
 } as Meta<typeof Autocomplete>;
+
+type SWCharacter = {
+  name: string;
+  height: string;
+  mass: string;
+  birth_year: string;
+};
 
 const defaultProps = {
   ...input.defaultVariants,
@@ -119,7 +137,7 @@ const DynamicTemplate = ({color, variant, ...args}: AutocompleteProps<Animal>) =
   </Autocomplete>
 );
 
-const RequiredTemplate = ({color, variant, ...args}: AutocompleteProps) => {
+const FormTemplate = ({color, variant, ...args}: AutocompleteProps) => {
   return (
     <form
       className="w-full max-w-xs items-start flex flex-col gap-4"
@@ -129,7 +147,6 @@ const RequiredTemplate = ({color, variant, ...args}: AutocompleteProps) => {
       }}
     >
       <Autocomplete
-        isRequired
         color={color}
         label="Favorite Animal"
         name="favorite-animal"
@@ -142,6 +159,76 @@ const RequiredTemplate = ({color, variant, ...args}: AutocompleteProps) => {
         Submit
       </button>
     </form>
+  );
+};
+
+const FullyControlledTemplate = () => {
+  // Store Autocomplete input value, selected option, open state, and items
+  // in a state tracker
+  const [fieldState, setFieldState] = React.useState({
+    selectedKey: "",
+    inputValue: "",
+    items: animalsData,
+  });
+
+  // Implement custom filtering logic and control what items are
+  // available to the Autocomplete.
+  const {startsWith} = useFilter({sensitivity: "base"});
+
+  // Specify how each of the Autocomplete values should change when an
+  // option is selected from the list box
+  const onSelectionChange = (key) => {
+    // eslint-disable-next-line no-console
+    console.log(`onSelectionChange ${key}`);
+    setFieldState((prevState) => {
+      let selectedItem = prevState.items.find((option) => option.value === key);
+
+      return {
+        inputValue: selectedItem?.label || "",
+        selectedKey: key,
+        items: animalsData.filter((item) => startsWith(item.label, selectedItem?.label || "")),
+      };
+    });
+  };
+
+  // Specify how each of the Autocomplete values should change when the input
+  // field is altered by the user
+  const onInputChange = (value) => {
+    // eslint-disable-next-line no-console
+    console.log(`onInputChange ${value}`);
+    setFieldState((prevState: any) => ({
+      inputValue: value,
+      selectedKey: value === "" ? null : prevState.selectedKey,
+      items: animalsData.filter((item) => startsWith(item.label, value)),
+    }));
+  };
+
+  // Show entire list if user opens the menu manually
+  const onOpenChange = (isOpen, menuTrigger) => {
+    if (menuTrigger === "manual" && isOpen) {
+      setFieldState((prevState) => ({
+        inputValue: prevState.inputValue,
+        selectedKey: prevState.selectedKey,
+        items: animalsData,
+      }));
+    }
+  };
+
+  return (
+    <Autocomplete
+      className="max-w-xs"
+      inputValue={fieldState.inputValue}
+      items={fieldState.items}
+      label="Favorite Animal"
+      placeholder="Search an animal"
+      selectedKey={fieldState.selectedKey}
+      variant="bordered"
+      onInputChange={onInputChange}
+      onOpenChange={onOpenChange}
+      onSelectionChange={onSelectionChange}
+    >
+      {(item) => <AutocompleteItem key={item.value}>{item.label}</AutocompleteItem>}
+    </Autocomplete>
   );
 };
 
@@ -233,6 +320,40 @@ const LabelPlacementTemplate = ({color, variant, ...args}: AutocompleteProps) =>
     </div>
   </div>
 );
+
+const AsyncFilteringTemplate = ({color, variant, ...args}: AutocompleteProps<SWCharacter>) => {
+  let list = useAsyncList<SWCharacter>({
+    async load({signal, filterText}) {
+      let res = await fetch(`https://swapi.py4e.com/api/people/?search=${filterText}`, {signal});
+      let json = await res.json();
+
+      return {
+        items: json.results,
+      };
+    },
+  });
+
+  return (
+    <Autocomplete
+      className="max-w-xs"
+      color={color}
+      inputValue={list.filterText}
+      isLoading={list.isLoading}
+      items={list.items}
+      label="Select a character"
+      placeholder="Type to search..."
+      variant={variant}
+      onInputChange={list.setFilterText}
+      {...args}
+    >
+      {(item) => (
+        <AutocompleteItem key={item.name} className="capitalize">
+          {item.name}
+        </AutocompleteItem>
+      )}
+    </Autocomplete>
+  );
+};
 
 const AsyncLoadingTemplate = ({color, variant, ...args}: AutocompleteProps<Pokemon>) => {
   const [isOpen, setIsOpen] = React.useState(false);
@@ -381,9 +502,9 @@ const ItemStartContentTemplate = ({color, variant, ...args}: AutocompleteProps<A
 );
 
 const ControlledTemplate = ({color, variant, ...args}: AutocompleteProps<Animal>) => {
-  const [value, setValue] = React.useState<Key>("cat");
+  const [value, setValue] = React.useState<Key | null>("cat");
 
-  const handleSelectionChange = (key: Key) => {
+  const handleSelectionChange = (key: Key | null) => {
     setValue(key);
   };
 
@@ -528,7 +649,7 @@ const CustomStylesTemplate = ({color, variant, ...args}: AutocompleteProps<User>
     <Autocomplete
       className="max-w-xs"
       classNames={{
-        base: "min-h-unit-16",
+        base: "min-h-16",
         listboxWrapper: "max-h-[400px]",
       }}
       color={color}
@@ -643,6 +764,45 @@ const CustomStylesWithCustomItemsTemplate = ({color, ...args}: AutocompleteProps
   );
 };
 
+const WithReactHookFormTemplate = (args: AutocompleteProps) => {
+  const {
+    register,
+    formState: {errors},
+    handleSubmit,
+  } = useForm({
+    defaultValues: {
+      withDefaultValue: "cat",
+      withoutDefaultValue: "",
+      requiredField: "",
+    },
+  });
+
+  const onSubmit = (data: any) => {
+    // eslint-disable-next-line no-console
+    console.log(data);
+    alert("Submitted value: " + JSON.stringify(data));
+  };
+
+  return (
+    <form className="flex w-full max-w-xs flex-col gap-2" onSubmit={handleSubmit(onSubmit)}>
+      <Autocomplete {...args} {...register("withDefaultValue")}>
+        {items}
+      </Autocomplete>
+      <Autocomplete {...args} {...register("withoutDefaultValue")}>
+        {items}
+      </Autocomplete>
+      <Autocomplete {...args} {...register("requiredField", {required: true})}>
+        {items}
+      </Autocomplete>
+
+      {errors.requiredField && <span className="text-danger">This field is required</span>}
+      <button className={button({class: "w-fit"})} type="submit">
+        Submit
+      </button>
+    </form>
+  );
+};
+
 export const Default = {
   render: Template,
   args: {
@@ -652,10 +812,11 @@ export const Default = {
 };
 
 export const Required = {
-  render: RequiredTemplate,
+  render: FormTemplate,
 
   args: {
     ...defaultProps,
+    isRequired: true,
   },
 };
 
@@ -689,17 +850,16 @@ export const DisabledOptions = {
   },
 };
 
-export const WithDescription = {
-  render: MirrorTemplate,
+export const LabelPlacement = {
+  render: LabelPlacementTemplate,
 
   args: {
     ...defaultProps,
-    description: "Select your favorite animal",
   },
 };
 
-export const LabelPlacement = {
-  render: LabelPlacementTemplate,
+export const AsyncFiltering = {
+  render: AsyncFilteringTemplate,
 
   args: {
     ...defaultProps,
@@ -727,6 +887,27 @@ export const EndContent = {
 
   args: {
     ...defaultProps,
+  },
+};
+
+export const IsInvalid = {
+  render: Template,
+
+  args: {
+    ...defaultProps,
+    isInvalid: true,
+    variant: "bordered",
+    defaultSelectedKey: "dog",
+    errorMessage: "Please select a valid animal",
+  },
+};
+
+export const WithDescription = {
+  render: MirrorTemplate,
+
+  args: {
+    ...defaultProps,
+    description: "Select your favorite animal",
   },
 };
 
@@ -762,45 +943,38 @@ export const WithErrorMessage = {
 
   args: {
     ...defaultProps,
+    isInvalid: true,
     errorMessage: "Please select an animal",
   },
 };
 
-export const IsInvalid = {
-  render: Template,
+export const WithErrorMessageFunction = {
+  render: FormTemplate,
 
   args: {
     ...defaultProps,
-    isInvalid: true,
-    variant: "bordered",
-    defaultSelectedKey: "dog",
-    errorMessage: "Please select a valid animal",
+    isRequired: true,
+    errorMessage: (value: ValidationResult) => {
+      if (value.validationDetails.valueMissing) {
+        return "Value is required";
+      }
+    },
   },
 };
 
-export const Controlled = {
-  render: ControlledTemplate,
+export const WithValidation = {
+  render: FormTemplate,
 
   args: {
     ...defaultProps,
-  },
-};
+    label: "Select Cat or Dog",
+    validate: (value) => {
+      if (value.selectedKey == null || value.selectedKey === "cat" || value.selectedKey === "dog") {
+        return;
+      }
 
-export const CustomSelectorIcon = {
-  render: Template,
-
-  args: {
-    ...defaultProps,
-    disableSelectorIconRotation: true,
-    selectorIcon: <SelectorIcon />,
-  },
-};
-
-export const CustomItems = {
-  render: CustomItemsTemplate,
-
-  args: {
-    ...defaultProps,
+      return "Please select a valid animal";
+    },
   },
 };
 
@@ -830,6 +1004,40 @@ export const WithAriaLabel = {
   },
 };
 
+export const WithReactHookForm = {
+  render: WithReactHookFormTemplate,
+
+  args: {
+    ...defaultProps,
+  },
+};
+
+export const Controlled = {
+  render: ControlledTemplate,
+
+  args: {
+    ...defaultProps,
+  },
+};
+
+export const CustomSelectorIcon = {
+  render: Template,
+
+  args: {
+    ...defaultProps,
+    disableSelectorIconRotation: true,
+    selectorIcon: <SelectorIcon />,
+  },
+};
+
+export const CustomItems = {
+  render: CustomItemsTemplate,
+
+  args: {
+    ...defaultProps,
+  },
+};
+
 export const CustomStyles = {
   render: CustomStylesTemplate,
 
@@ -842,6 +1050,13 @@ export const CustomStyles = {
 export const CustomStylesWithCustomItems = {
   render: CustomStylesWithCustomItemsTemplate,
 
+  args: {
+    ...defaultProps,
+  },
+};
+
+export const FullyControlled = {
+  render: FullyControlledTemplate,
   args: {
     ...defaultProps,
   },