+++ packages/core/theme/src/utils/index.ts
@@ -1,3 +1,14 @@
-export * from "./classes";
-export * from "./types";
-export * from "./variants";
+export {
+  baseStyles,
+  ringClasses,
+  focusVisibleClasses,
+  dataFocusVisibleClasses,
+  groupDataFocusVisibleClasses,
+  translateCenterClasses,
+  absoluteFullClasses,
+  collapseAdjacentVariantBorders,
+} from "./classes";
+export type {SlotsToClasses} from "./types";
+export {colorVariants} from "./variants";
+export {COMMON_UNITS, twMergeConfig} from "./tw-merge-config";
+export {cn} from "./cn";